change count for GRES and TRES to uint64

+++ daemon.go
@@ -942,8 +942,8 @@ func ccSyncStats() error {
 				gresTotal, errTotal := SlurmParseGRES(*stat.Gres.Total)
 				gresAlloc, errAlloc := SlurmParseGRES(*stat.Gres.Used)
 				if errTotal == nil && errAlloc == nil {
-					node.GpusTotal = gresTotal.Count
-					node.GpusAllocated = gresAlloc.Count
+					node.GpusTotal = int(gresTotal.Count)
+					node.GpusAllocated = int(gresAlloc.Count)
 				} else {
 					node.GpusTotal = 0
 					node.GpusAllocated = 0
@@ -1082,11 +1082,11 @@ func slurmJobToCcStartJob(job SacctJob) (*StartJob, error) {
 	setResources := func(tresList []SacctJobTres, ccStartJob *StartJob) {
 		for _, tres := range tresList {
 			if *tres.Type == "cpu" {
-				ccStartJob.NumHWThreads = *tres.Count
+				ccStartJob.NumHWThreads = int32(*tres.Count)
 			}
 
 			if *tres.Type == "gres" && *tres.Name == "gpu" {
-				ccStartJob.NumAcc = *tres.Count
+				ccStartJob.NumAcc = int32(*tres.Count)
 			}
 		}
 	}
